Rename AddressEditor to GranularAddressEditor
