+++ src/Place/ReadModel/RDF/RdfProjector.php
@@ -6,16 +6,14 @@ namespace CultuurNet\UDB3\Place\ReadModel\RDF;
 
 use Broadway\Domain\DomainMessage;
 use Broadway\EventHandling\EventListener;
-use CultuurNet\UDB3\Address\Address as LegacyAddress;
 use CultuurNet\UDB3\Address\AddressParser;
-use CultuurNet\UDB3\Address\FullAddressFormatter;
 use CultuurNet\UDB3\Geocoding\Coordinate\Coordinates;
 use CultuurNet\UDB3\Iri\IriGeneratorInterface;
 use CultuurNet\UDB3\Model\Place\ImmutablePlace;
 use CultuurNet\UDB3\Model\Place\Place;
-use CultuurNet\UDB3\Model\ValueObject\Geography\TranslatedAddress;
 use CultuurNet\UDB3\Model\ValueObject\Text\TranslatedTitle;
 use CultuurNet\UDB3\Place\Events\PlaceProjectedToJSONLD;
+use CultuurNet\UDB3\RDF\Editor\AddressEditor;
 use CultuurNet\UDB3\RDF\Editor\GraphEditor;
 use CultuurNet\UDB3\RDF\Editor\WorkflowStatusEditor;
 use CultuurNet\UDB3\RDF\GraphRepository;
@@ -35,19 +33,10 @@ final class RdfProjector implements EventListener
     private AddressParser $addressParser;
 
     private const TYPE_LOCATIE = 'dcterms:Location';
-    private const TYPE_ADRES = 'locn:Address';
     private const TYPE_GEOMETRIE = 'locn:Geometry';
 
     private const PROPERTY_LOCATIE_NAAM = 'locn:locatorName';
-    private const PROPERTY_LOCATIE_ADRES = 'locn:address';
     private const PROPERTY_LOCATIE_GEOMETRIE = 'locn:geometry';
-
-    private const PROPERTY_ADRES_STRAATNAAM = 'locn:thoroughfare';
-    private const PROPERTY_ADRES_HUISNUMMER = 'locn:locatorDesignator';
-    private const PROPERTY_ADRES_POSTCODE = 'locn:postcode';
-    private const PROPERTY_ADRES_GEMEENTENAAM = 'locn:postName';
-    private const PROPERTY_ADRES_LAND = 'locn:adminUnitL1';
-    private const PROPERTY_ADRES_VOLLEDIG_ADRES = 'locn:fullAddress';
 
     private const PROPERTY_GEOMETRIE_GML = 'geosparql:asGML';
 
@@ -90,7 +79,7 @@ final class RdfProjector implements EventListener
 
         $this->setTitle($resource, $place->getTitle());
 
-        $this->setAddress($resource, $place->getAddress());
+        (new AddressEditor($this->addressParser))->setAddress($resource, $place->getAddress());
 
         if ($place->getGeoCoordinates()) {
             $this->setCoordinates($resource, $place->getGeoCoordinates());
@@ -117,54 +106,6 @@ final class RdfProjector implements EventListener
                 self::PROPERTY_LOCATIE_NAAM,
                 new Literal($translatedTitle->getTranslation($language)->toString(), $language->toString())
             );
-        }
-    }
-
-    private function setAddress(Resource $resource, TranslatedAddress $translatedAddress): void
-    {
-        foreach ($translatedAddress->getLanguages() as $language) {
-            $address = $translatedAddress->getTranslation($language);
-
-            if (!$resource->hasProperty(self::PROPERTY_LOCATIE_ADRES)) {
-                $resource->add(self::PROPERTY_LOCATIE_ADRES, $resource->getGraph()->newBNode([self::TYPE_ADRES]));
-            }
-            $addressResource = $resource->getResource(self::PROPERTY_LOCATIE_ADRES);
-
-            $countryCode = $address->getCountryCode()->toString();
-            if ($addressResource->get(self::PROPERTY_ADRES_LAND) !== $countryCode) {
-                $addressResource->set(self::PROPERTY_ADRES_LAND, $countryCode);
-            }
-
-            $postalCode = $address->getPostalCode()->toString();
-            if ($addressResource->get(self::PROPERTY_ADRES_POSTCODE) !== $postalCode) {
-                $addressResource->set(self::PROPERTY_ADRES_POSTCODE, $postalCode);
-            }
-
-            $addressFormatter = new FullAddressFormatter();
-            $formattedAddress = $addressFormatter->format(LegacyAddress::fromUdb3ModelAddress($address));
-            $parsedAddress = $this->addressParser->parse($formattedAddress);
-
-            $houseNumber = $parsedAddress ? $parsedAddress->getHouseNumber() : null;
-            if ($houseNumber !== null) {
-                $addressResource->set(self::PROPERTY_ADRES_HUISNUMMER, $houseNumber);
-            }
-
-            $addressResource->addLiteral(
-                self::PROPERTY_ADRES_VOLLEDIG_ADRES,
-                new Literal($formattedAddress, $language->toString())
-            );
-
-            $addressResource->addLiteral(
-                self::PROPERTY_ADRES_GEMEENTENAAM,
-                new Literal($address->getLocality()->toString(), $language->toString())
-            );
-
-            if ($parsedAddress && $parsedAddress->getThoroughfare() !== null) {
-                $addressResource->addLiteral(
-                    self::PROPERTY_ADRES_STRAATNAAM,
-                    new Literal($parsedAddress->getThoroughfare(), $language->toString())
-                );
-            }
         }
     }
