+++ src/RDF/Editor/AddressEditor.php
@@ -5,22 +5,19 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\RDF\Editor;
 
 use CultuurNet\UDB3\Address\Address as LegacyAddress;
-use CultuurNet\UDB3\Address\AddressFormatter;
 use CultuurNet\UDB3\Address\AddressParser;
 use CultuurNet\UDB3\Address\FullAddressFormatter;
-use CultuurNet\UDB3\Model\ValueObject\Geography\Address;
-use CultuurNet\UDB3\Model\ValueObject\Geography\Locality;
-use CultuurNet\UDB3\Model\ValueObject\Geography\PostalCode;
-use CultuurNet\UDB3\Model\ValueObject\Geography\Street;
-use CultuurNet\UDB3\Model\ValueObject\Translation\Language;
-use CultuurNet\UDB3\RDF\MainLanguageRepository;
-use EasyRdf\Graph;
+use CultuurNet\UDB3\Model\ValueObject\Geography\TranslatedAddress;
+use EasyRdf\Literal;
 use EasyRdf\Resource;
 
 final class AddressEditor
 {
+    private AddressParser $addressParser;
+
     private const TYPE_ADRES = 'locn:Address';
 
+    private const PROPERTY_LOCATIE_ADRES = 'locn:address';
     private const PROPERTY_ADRES_STRAATNAAM = 'locn:thoroughfare';
     private const PROPERTY_ADRES_HUISNUMMER = 'locn:locatorDesignator';
     private const PROPERTY_ADRES_POSTCODE = 'locn:postcode';
@@ -28,161 +25,57 @@ final class AddressEditor
     private const PROPERTY_ADRES_LAND = 'locn:adminUnitL1';
     private const PROPERTY_ADRES_VOLLEDIG_ADRES = 'locn:fullAddress';
 
-    private Graph $graph;
-    private MainLanguageRepository $mainLanguageRepository;
-    private AddressParser $addressParser;
-    private AddressFormatter $addressFormatter;
-
-    private function __construct(
-        Graph $graph,
-        MainLanguageRepository $mainLanguageRepository,
-        AddressParser $addressParser
-    ) {
-        $this->graph = $graph;
-        $this->mainLanguageRepository = $mainLanguageRepository;
+    public function __construct(AddressParser $addressParser)
+    {
         $this->addressParser = $addressParser;
-        $this->addressFormatter = new FullAddressFormatter();
     }
 
-    public static function for(
-        Graph $graph,
-        MainLanguageRepository $mainLanguageRepository,
-        AddressParser $addressParser
-    ): self {
-        return new self($graph, $mainLanguageRepository, $addressParser);
-    }
-
-    public function addAddress(
-        string $resourceIri,
-        Address $address,
-        string $property
-    ): void {
-        $resource = $this->graph->resource($resourceIri);
-
-        if (!$resource->hasProperty($property)) {
-            $resource->add($property, $resource->getGraph()->newBNode());
-        }
-
-        $addressResource = $resource->getResource($property);
-        if ($addressResource->type() !== self::TYPE_ADRES) {
-            $addressResource->setType(self::TYPE_ADRES);
-        }
-
-        $countryCode = $address->getCountryCode()->toString();
-        if ($addressResource->get(self::PROPERTY_ADRES_LAND) !== $countryCode) {
-            $addressResource->set(self::PROPERTY_ADRES_LAND, $countryCode);
-        }
-
-        $postalCode = $address->getPostalCode()->toString();
-        if ($addressResource->get(self::PROPERTY_ADRES_POSTCODE) !== $postalCode) {
-            $addressResource->set(self::PROPERTY_ADRES_POSTCODE, $postalCode);
-        }
-
-        $parsedAddress = $this->addressParser->parse(
-            $this->addressFormatter->format(LegacyAddress::fromUdb3ModelAddress($address))
-        );
-        $houseNumber = $parsedAddress ? $parsedAddress->getHouseNumber() : null;
-        if ($houseNumber !== null) {
-            $addressResource->set(self::PROPERTY_ADRES_HUISNUMMER, $houseNumber);
-        }
-
-        $mainLanguage = $this->mainLanguageRepository->get($resourceIri, new Language('nl'))->toString();
-        $this->updateTranslatableAddress($resourceIri, $address, $mainLanguage, $property);
-    }
-
-    public function removeAddresses(): void
+    public function setAddress(Resource $resource, TranslatedAddress $translatedAddress): void
     {
-        $resources = $this->graph->allOfType(self::TYPE_ADRES);
-        foreach ($resources as $resource) {
-            $resource->delete('rdf:type');
+        foreach ($translatedAddress->getLanguages() as $language) {
+            $address = $translatedAddress->getTranslation($language);
 
-            $resource->set(self::PROPERTY_ADRES_LAND, null);
-            $resource->set(self::PROPERTY_ADRES_POSTCODE, null);
-            $resource->set(self::PROPERTY_ADRES_STRAATNAAM, null);
-            $resource->set(self::PROPERTY_ADRES_HUISNUMMER, null);
-            $resource->set(self::PROPERTY_ADRES_VOLLEDIG_ADRES, null);
-            $resource->set(self::PROPERTY_ADRES_GEMEENTENAAM, null);
-        }
-    }
+            if (!$resource->hasProperty(self::PROPERTY_LOCATIE_ADRES)) {
+                $resource->add(self::PROPERTY_LOCATIE_ADRES, $resource->getGraph()->newBNode([self::TYPE_ADRES]));
+            }
+            $addressResource = $resource->getResource(self::PROPERTY_LOCATIE_ADRES);
 
-    public function getAddress(): ?Resource
-    {
-        $resources = $this->graph->allOfType(self::TYPE_ADRES);
+            $countryCode = $address->getCountryCode()->toString();
+            if ($addressResource->get(self::PROPERTY_ADRES_LAND) !== $countryCode) {
+                $addressResource->set(self::PROPERTY_ADRES_LAND, $countryCode);
+            }
 
-        if (count($resources) === 0) {
-            return null;
-        }
+            $postalCode = $address->getPostalCode()->toString();
+            if ($addressResource->get(self::PROPERTY_ADRES_POSTCODE) !== $postalCode) {
+                $addressResource->set(self::PROPERTY_ADRES_POSTCODE, $postalCode);
+            }
 
-        return $resources[0];
-    }
+            $addressFormatter = new FullAddressFormatter();
+            $formattedAddress = $addressFormatter->format(LegacyAddress::fromUdb3ModelAddress($address));
+            $parsedAddress = $this->addressParser->parse($formattedAddress);
 
-    public function updateTranslatableAddress(
-        string $resourceIri,
-        Address $address,
-        string $language,
-        string $property
-    ): void {
-        $resource = $this->graph->resource($resourceIri);
+            $houseNumber = $parsedAddress ? $parsedAddress->getHouseNumber() : null;
+            if ($houseNumber !== null) {
+                $addressResource->set(self::PROPERTY_ADRES_HUISNUMMER, $houseNumber);
+            }
 
-        /** @var Resource|null $addressResource */
-        $addressResource = $resource->getResource($property);
-        if ($addressResource === null) {
-            // This is a case that should not happen in reality, since every new place should get a locn:Address via
-            // handleAddressUpdated().
-            return;
-        }
-
-        // The locn:fullAddress predicate is set per language since it contains language-specific info like the street
-        // name and municipality name. It is included because not all addresses can be parsed into the expected
-        // thoroughfare and house number, so in those cases at least the full address is completed and consumers can
-        // always try to parse it themselves if wanted.
-        $graphEditor = GraphEditor::for($addressResource->getGraph())->replaceLanguageValue(
-            $addressResource->getUri(),
-            self::PROPERTY_ADRES_VOLLEDIG_ADRES,
-            $this->addressFormatter->format(LegacyAddress::fromUdb3ModelAddress($address)),
-            $language
-        );
-
-        // Always set the locn:postName predicate based on the Address, not the ParsedAddress, because in some cases an
-        // address cannot be parsed (e.g. it's outside of Belgium, or the street address could not be parsed/found), but
-        // the original address always contains the right municipality in any case.
-        $graphEditor->replaceLanguageValue(
-            $addressResource->getUri(),
-            self::PROPERTY_ADRES_GEMEENTENAAM,
-            $address->getLocality()->toString(),
-            $language
-        );
-
-        $parsedAddress = $this->addressParser->parse(
-            $this->addressFormatter->format(LegacyAddress::fromUdb3ModelAddress($address))
-        );
-        // Only set the locn:thoroughfare predicate based on the ParsedAddress (if given), not the street in the
-        // original Address, because locn:thoroughfare MUST NOT contain a house number. If there is no ParsedAddress
-        // remove the value for the given language instead since it will probably be outdated (if set previously).
-        // Keep in mind that locn:thoroughfare is optional.
-        if ($parsedAddress && $parsedAddress->getThoroughfare() !== null) {
-            $graphEditor->replaceLanguageValue(
-                $addressResource->getUri(),
-                self::PROPERTY_ADRES_STRAATNAAM,
-                $parsedAddress->getThoroughfare(),
-                $language
+            $addressResource->addLiteral(
+                self::PROPERTY_ADRES_VOLLEDIG_ADRES,
+                new Literal($formattedAddress, $language->toString())
             );
-        } else {
-            $graphEditor->deleteLanguageValue(
-                $addressResource->getUri(),
-                self::PROPERTY_ADRES_STRAATNAAM,
-                $language
+
+            $addressResource->addLiteral(
+                self::PROPERTY_ADRES_GEMEENTENAAM,
+                new Literal($address->getLocality()->toString(), $language->toString())
             );
+
+            if ($parsedAddress && $parsedAddress->getThoroughfare() !== null) {
+                $addressResource->addLiteral(
+                    self::PROPERTY_ADRES_STRAATNAAM,
+                    new Literal($parsedAddress->getThoroughfare(), $language->toString())
+                );
+            }
         }
     }
 
-    public static function fromLegacyAddress(LegacyAddress $legacyAddress): Address
-    {
-        return new Address(
-            new Street($legacyAddress->getStreetAddress()->toNative()),
-            new PostalCode($legacyAddress->getPostalCode()->toNative()),
-            new Locality($legacyAddress->getLocality()->toNative()),
-            $legacyAddress->getCountryCode()
-        );
-    }
 }
